Create firstScriptDirectFileAccess.scpt
All those switches to TextEdit are not necessary, because files can be written with inbuilt AppleScript functions very easily. I did a mixture between copy/paste and writing to file in a very dirty and quick approach. So if you like, have a look at this and make changes as appropriate.